backend/agents/backend_agent/requirements.txt
google-generativeai==0.3.2
flask==2.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
pydantic==2.5.2
docker==6.1.3  # For container validation